Hackathon_back/requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-dotenv
pydantic
langchain_cohere
langchain_core

